+++ core/src/main/kotlin/supplier/RestEntitySupplier.kt
@@ -4,6 +4,7 @@ import dev.kord.common.entity.DiscordAuditLogEntry
 import dev.kord.common.entity.Snowflake
 import dev.kord.common.entity.optional.OptionalSnowflake
 import dev.kord.common.entity.optional.optionalSnowflake
+import dev.kord.common.exception.RequestException
 import dev.kord.core.*
 import dev.kord.core.cache.data.*
 import dev.kord.core.entity.*
@@ -105,12 +106,23 @@ public class RestEntitySupplier(public val kord: Kord) : EntitySupplier {
 
     /**
      * Returns the onboarding for a specified [guildId], or null if one cannot be found.
+     *
      * @param guildId The ID of the guild to get the onboarding for.
+     * @throws RequestException if something went wrong during the request.
      */
     public suspend fun getGuildOnboardingOrNull(guildId: Snowflake): GuildOnboarding? = catchNotFound {
         val onboarding = guild.getGuildOnboarding(guildId)
         GuildOnboarding(onboarding, kord)
     }
+
+    /**
+     * Returns the onboarding for a specified [guildId].
+     * @param guildId The ID of the guild to get the onboarding for.
+     * @throws RequestException if something went wrong during the request.
+     * @throws EntityNotFoundException if the preview was not found.
+     */
+    public suspend fun getGuildOnboarding(guildId: Snowflake): GuildOnboarding =
+            GuildOnboarding(guild.getGuildOnboarding(guildId), kord)
 
     override suspend fun getMemberOrNull(guildId: Snowflake, userId: Snowflake): Member? = catchNotFound {
         val member = guild.getGuildMember(guildId = guildId, userId = userId)
